app/requirements.txt
Babel==2.6.0
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.15
django-debug-toolbar==1.9.1
django-extensions==2.1.0
django-extra-fields==1.0.0
django-filter==2.0.0
django-markdownx==2.0.23
django-phonenumber-field==2.0.1
django-rest-swagger==2.2.0
djangorestframework==3.8.2
djangorestframework-filters==0.10.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
nose==1.3.7
numpy==1.15.0
openapi-codec==1.3.2
phonenumbers==8.9.10
Pillow==5.2.0
psycopg2==2.7.5
PyJWT==1.6.4
python-decouple==3.1
pytz==2018.5
qrcode==6.0
requests==2.19.1
simplejson==3.16.0
six==1.11.0
sqlparse==0.2.4
typing==3.6.4
uritemplate==3.0.0
urllib3==1.23
